+++ packages/pob-dependencies/lib/index.d.ts
@@ -1,7 +1,7 @@
 interface Dependencies {
-  '@babel/core': '7.12.9';
-  '@babel/preset-env': '7.12.7';
-  '@babel/preset-react': '7.12.7';
+  '@babel/core': '7.12.10';
+  '@babel/preset-env': '7.12.10';
+  '@babel/preset-react': '7.12.10';
   '@babel/runtime': '7.12.5';
   '@chrp/typedoc-plugin-lerna-packages': '0.4.1';
   '@pob/commitlint-config': '^2.0.1';
@@ -11,11 +11,11 @@ interface Dependencies {
   '@pob/eslint-config-typescript-node': '41.0.4';
   '@pob/eslint-config-typescript-react': '41.0.4';
   '@pob/lerna-light': '^3.23.0';
-  '@pob/pretty-pkg': '^2.0.1';
-  '@pob/root': '^3.0.2';
-  '@types/jest': '26.0.16';
-  '@typescript-eslint/eslint-plugin': '4.9.0';
-  '@typescript-eslint/parser': '4.9.0';
+  '@pob/pretty-pkg': '^2.0.2';
+  '@pob/root': '^3.0.3';
+  '@types/jest': '26.0.19';
+  '@typescript-eslint/eslint-plugin': '4.9.1';
+  '@typescript-eslint/parser': '4.9.1';
   'babel-jest': '26.6.3';
   'babel-preset-latest-node': '5.2.0';
   'babel-preset-modern-browsers': '15.0.2';
@@ -27,18 +27,18 @@ interface Dependencies {
   'eslint-plugin-react': '7.21.5';
   'eslint-plugin-react-hooks': '4.2.0';
   'eslint-plugin-unicorn': '23.0.0';
-  husky: '4.3.4';
+  husky: '4.3.5';
   jest: '26.6.3';
   'jest-junit-reporter': '1.1.0';
-  'pob-babel': '^26.0.1';
+  'pob-babel': '^26.1.0';
   'pob-lcov-reporter': '^4.0.1';
   prettier: '2.2.1';
   'repository-check-dirty': '^2.0.1';
-  rollup: '2.34.1';
+  rollup: '2.34.2';
   semver: '7.3.4';
   'standard-version': '9.0.0';
   typedoc: '0.19.2';
-  typescript: '4.1.2';
+  typescript: '4.1.3';
 }
 
 declare const dependencies: Dependencies;